Update jar locations
For the FFT methods to work in Jython scripts launched in the IDE,
define the path variable MVN_LOCAL_REPO in your workspace's Linked
Resources to, say, $HOME/.m2/repository or where localRepository is set
in $HOME/.m2/settings.xml. Also tweak some NumPy type
aliases removed from version 2, add import error messaging and a None
check

+++ uk.ac.diamond.scisoft.python/src/scisoftpy/external.py
@@ -336,10 +336,11 @@ def get_python(py3=True):
         if exe is None:
             raise RuntimeError('Problem with running external process: ' + p.stderr.read())
     finally:
-        p.stdout.close()
-        p.stderr.close()
-        p.terminate()
-        p.wait()
+        if p:
+            p.stdout.close()
+            p.stderr.close()
+            p.terminate()
+            p.wait()
     return exe, path, ldpath
 
 def parse_for_env(stream, sep=':'):
